+++ src/app/problems/_forms/parsons-form.ts
@@ -1,5 +1,5 @@
 import {FormBuilder, FormControl, FormGroup, Validators} from "@angular/forms";
-import {ParsonsFormData} from "@app/problems/_models/parsons/parsons-form-data";
+import {Question} from "@app/_models";
 
 export class ParsonsForm {
     /**
@@ -21,27 +21,55 @@ export class ParsonsForm {
 
     /**
      * Creates a FormGroup for a Parsons question with existing data.
-     * @param data - The existing data for the question.
+     * @param question - The question object.
+     * @param event - The event.
+     * @param course - The course.
      */
-    static createFormWithData(data: ParsonsFormData): FormGroup {
+    static createFormWithData(question: Question, event: number, course: number): FormGroup {
         const builder = new FormBuilder();
         return builder.group({
-            title: new FormControl(data.title, [Validators.required]),
-            difficulty: new FormControl(data.difficulty, [Validators.required]),
-            category: new FormControl(data.category),
-            course: new FormControl(data.course),
-            event: new FormControl(data.event),
-            junit_template: new FormControl(data.junit_template),
-            lines: new FormControl(data.lines, [Validators.required]),
-            additional_file_name: new FormControl(data.additional_file_name),
+            title: new FormControl(question.title, [Validators.required]),
+            difficulty: new FormControl(question.difficulty, [Validators.required]),
+            category: new FormControl(question.category),
+            course: new FormControl(course),
+            event: new FormControl(event),
+            junit_template: new FormControl(question.junit_template),
+            lines: new FormControl(question.lines.join('\n'), [Validators.required]),
+            additional_file_name: new FormControl(question.additional_file_name),
         });
     }
 
     /**
      * Extracts the data from the FormGroup.
      * @param form - The FormGroup for the question.
+     * @param variablesJSON - The variables in JSON format.
+     * @param questionText - The question text.
      */
-    static extractData(form: FormGroup): ParsonsFormData {
-        return form.value;
+    static extractData(form: FormGroup, variablesJSON: JSON[], questionText: string): ParsonsFormData {
+        return {
+            title: form.value.title,
+            difficulty: form.value.difficulty,
+            course: form.value.course,
+            event: form.value.event,
+            text: questionText,
+            category: form.value.category,
+            variables: variablesJSON,
+            lines: form.value.lines.split('\n'),
+            additional_file_name: form.value.additional_file_name,
+            junit_template: form.value.junit_template,
+        };
     }
+}
+
+export interface ParsonsFormData {
+    title: string,
+    difficulty: string,
+    course: number,
+    event: number,
+    text: string,
+    category: number,
+    variables: JSON[],
+    lines: string[],
+    additional_file_name: string,
+    junit_template: string,
 }
