+++ src/app/problems/_forms/problem-set-form.ts
@@ -1,5 +1,4 @@
 import {FormBuilder, FormControl, FormGroup} from "@angular/forms";
-import {JavaFormData} from "@app/problems/_models/java/java-form-data";
 
 export class ProblemSetForm {
     /**
@@ -20,7 +19,18 @@ export class ProblemSetForm {
      * Extracts the data from the FormGroup.
      * @param form - The FormGroup for the problem set.
      */
-    static extractData(form: FormGroup): JavaFormData {
+    static extractData(form: FormGroup): ProblemSetFormData {
         return form.value;
     }
+}
+
+export interface ProblemSetFormData {
+    page: number,
+    page_size: number,
+    search: string,
+    parentCategory: string,
+    subCategory: string,
+    difficulty: string,
+    is_sample: string,
+    ordering: string,
 }
