--- src/app/problems/_models/java/java-form-data.ts
@@ -1,8 +1,0 @@
-export interface JavaFormData {
-    title: string,
-    difficulty: string,
-    course: number,
-    event: number,
-    category: number,
-    junit_template: string,
-}
