--- src/app/problems/_models/java/java-request-data.ts
@@ -1,11 +1,0 @@
-export interface JavaRequestData {
-    title: string,
-    difficulty: string,
-    course: number,
-    event: number,
-    text: string,
-    category: number,
-    variables: JSON[],
-    junit_template: string,
-    input_file_names: JSON,
-}
