--- src/app/problems/_models/mcq/mcq-form-data.ts
@@ -1,8 +1,0 @@
-export interface McqFormData {
-    title: string,
-    difficulty: string,
-    course: number,
-    event: number,
-    category: number,
-    visible_distractor_count: number
-}
