--- src/app/problems/_models/mcq/mcq-request-data.ts
@@ -1,12 +1,0 @@
-export interface McqRequestData {
-    title: string,
-    difficulty: string,
-    course: number,
-    event: number,
-    text: string,
-    answer: string,
-    category: number,
-    variables: JSON[],
-    visible_distractor_count: number,
-    choices: { [id: string]: string },
-}
