--- src/app/problems/_models/parsons/parsons-form-data.ts
@@ -1,10 +1,0 @@
-export interface ParsonsFormData {
-    title: string,
-    difficulty: string,
-    course: number,
-    event: number,
-    category: number,
-    lines: string,
-    additional_file_name: string,
-    junit_template: string
-}
