--- src/app/problems/_models/parsons/parsons-request-data.ts
@@ -1,12 +1,0 @@
-export interface ParsonsRequestData {
-    title: string,
-    difficulty: string,
-    course: number,
-    event: number,
-    text: string,
-    category: number,
-    variables: JSON[],
-    lines: string[],
-    additional_file_name: string,
-    junit_template: string,
-}
