--- src/app/problems/_models/parsons/parsons-submission-request.ts
@@ -1,12 +1,0 @@
-export interface ParsonsSubmissionRequest {
-    title: string,
-    difficulty: string,
-    course: number,
-    event: number,
-    text: string,
-    category: number,
-    variables: JSON[],
-    junit_template: string,
-    lines: string[],
-    additional_file_name: string,
-}
