--- src/app/problems/_models/problem-set-options.ts
@@ -1,10 +1,0 @@
-export interface ProblemSetOptions {
-    page: number,
-    page_size: number,
-    search: string,
-    parentCategory: string,
-    subCategory: string,
-    difficulty: string,
-    is_sample: string,
-    ordering: string,
-}
