+++ src/app/problems/problem-create/java-create-snippet/java-create-snippet.component.ts
@@ -7,7 +7,6 @@ import {QuestionService} from '@app/_services/api/question.service';
 import {ToastrService} from "ngx-toastr";
 import {CourseService} from '@app/_services/api/course/course.service';
 import {CategoryService} from '@app/_services/api/category.service';
-import {ProblemHelpersService} from '@app/_services/problem-helpers.service';
 import {JavaForm} from "@app/problems/_forms/java-form";
 
 @Component({
@@ -30,8 +29,7 @@ export class JavaCreateSnippetComponent implements OnInit {
                 private formBuilder: FormBuilder,
                 private toastr: ToastrService,
                 private courseService: CourseService,
-                private categoryService: CategoryService,
-                private problemHelpersService: ProblemHelpersService) {
+                private categoryService: CategoryService) {
     }
 
     /**
@@ -80,14 +78,12 @@ export class JavaCreateSnippetComponent implements OnInit {
      * Form submission.
      */
     onSubmit(): void {
-        const data = JavaForm.extractData(this.formGroup);
-        const submissionRequest = this.problemHelpersService.createJavaSubmissionRequest(data, this.variables, this.inputFileNames, this.questionText);
+        const submissionRequest = JavaForm.extractData(this.formGroup, this.variables, this.inputFileNames, this.questionText);
         this.questionService.postJavaQuestion(submissionRequest)
-            .subscribe((result) => {
+            .subscribe(() => {
                 window.scroll(0, 0);
                 this.formGroup.reset();
-                if (result.success != false)
-                    this.toastr.success('The Question has been Created Successfully.');
+                this.toastr.success('The Question has been Created Successfully.');
             });
     }
 }
